app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.7
flasgger==0.9.5
SQLAlchemy==2.0.19
python-dotenv==1.0.0